\name{count_data}
\alias{count_data}
\docType{data}
\title{
Count Format Data Example
}
\description{A toy data set in count format.
}
\usage{count_data}
\format{
  A data frame with 3 rows on the following 5 variables.
  \describe{
    \item{\code{id}}{An identification variable}
    \item{\code{time}}{Aggregate time variable}
    \item{\code{event}}{Aggregated status indicator variable}
    \item{\code{var1}}{First example explanatory variable}
    \item{\code{var2}}{Second example explanatory variable}
  }
}

\examples{count_data
}
\keyword{datasets}
