\name{cp_data}
\alias{cp_data}
\docType{data}
\title{
Counting Process Data Example
}
\description{
A toy data set in counting process format.}
\usage{cp_data}
\format{
  A data frame with 6 rows on the following 6 variables.
  \describe{
    \item{\code{id}}{An identification variable}
    \item{\code{time1}}{Starting time of observation interval}
    \item{\code{time2}}{Ending time of observation interval}
    \item{\code{event}}{Status indicator variable}
    \item{\code{var1}}{First example explanatory variable}
    \item{\code{var2}}{Second example explanatory variable}
  }
}
\examples{
cp_data
}
\keyword{datasets}

