\name{long_data}
\alias{long_data}
\docType{data}
\title{
Long Format Data Example
}
\description{
A toy data set in long format data.
}
\usage{long_data}
\format{
  A data frame with 9 rows on the following 5 variables.
  \describe{
    \item{\code{id}}{An identification variable}
    \item{\code{time}}{Time of observation}
    \item{\code{event}}{Status indicator variable}
    \item{\code{var1}}{First example explanatory variable}
    \item{\code{var2}}{Second example explanatory variable}
  }
}
\examples{
long_data
}
\keyword{datasets}
