\name{wide_data}
\alias{wide_data}
\docType{data}
\title{Wide Format Data Example
}
\description{A toy data set in wide format.
}
\usage{wide_data}
\format{
  A data frame with 3 rows on the following 14 variables.
  \describe{
    \item{\code{id}}{An identification variable}
    \item{\code{time1}}{First time observation column}
    \item{\code{time2}}{Second time observation column}
    \item{\code{time3}}{Third time observation column}
    \item{\code{time4}}{Fourth observation column}
    \item{\code{event1}}{Status indicator at first time}
    \item{\code{event2}}{Status indicator at second time}
    \item{\code{event3}}{Status indicator at third time}
    \item{\code{event4}}{Status indicator at fourth time}
    \item{\code{var11}}{First explanatory variable at first time}
    \item{\code{var12}}{First explanatory variable at second time}
    \item{\code{var13}}{First explanatory variable at third time}
    \item{\code{var14}}{First explanatory variable at fourth time}
    \item{\code{var2}}{Second explanatory variable}
  }
}

\examples{
wide_data
}
\keyword{datasets}
