% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsd-package.R
\docType{package}
\name{wlsd-package}
\alias{wlsd}
\alias{wlsd-package}
\title{wlsd: Wrangling Longitudinal Survival Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Streamlines the process of transitioning between data formats commonly used in survival analysis. Functions convert longitudinal data between formats used as input for survival models as well as support overall preparation. Users are able to focus on model building rather than data wrangling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ci2131a/wlsd}
  \item Report bugs at \url{https://github.com/ci2131a/wlsd/issues}
}

}
\author{
\strong{Maintainer}: Charles Ingulli \email{charlesfi@outlook.com}

}
\keyword{internal}
