
\name{basedate}
\alias{basedate}
\title{Create Baseline Row}
\usage{
basedate(data,id)
}
\description{
Creates a new row of values for subjects representing baseline observations in a data set of follow-up observations.
}
\arguments{
  \item{data}{A data frame with relevant columns.}
  \item{id}{A character string of the identification column name in \code{data}.}
}
\examples{



}
\value{A data frame in longitudinal form.}
\details{

}

