
\name{LBP}
\alias{LBP}
\title{Low Back Pain Data Set}
\usage{
LBP
}
\description{
This is data collected by ? to investigate the risk of low back pain (LBP) in a cohort of manufaturing workers.
}
\format{
  \tabular{ll}{
    sid: \tab The subject identification variable for individuals. \cr
    Baseline.date: \tab The date of baseline visit or enrollment of individuals into the study. \cr
    Date: \tab The calendar time of follow up. \cr
    time_to_row: \tab The number of days between the current follow up period and the baseline date. \cr
    case.lbp: \tab A status indicator for individuals possessing LBP. \cr
    case.med: \tab A status indicator determining whether indviduals are prescribed medication for LBP. \cr
    case.sc: \tab A status indicator to determine whether individuals are seeking care for LBP. \cr
    case.ls: \tab A status indicator to determine whether individuals have lost time from work due to LBP. \cr
    gender: \tab The assigned sex of the individual (either M for Male or F for Female). \cr
    age: \tab The age of the individual at baseline visit. \cr
    weight: \tab The weight of individuals in ?. \cr
    height: \tab The height of individuals in ?. \cr
    raceth: \tab A categorical variable to determine the race/ethnicity of individuals. Possible levels are ... \cr
    smoking: \tab A smoking indicator variable. \cr
    comptenure: \tab A categorical variable to determine whether individuals have \cr
    jobtenure: \tab A categorical variable to determine whether individuals have tenure in their job. \cr
    control.order: \tab  A categorical variable to determine whether individuals have control over the order in which they complete tasks. \cr
    control.pace: \tab  A categorical variable to determine whether individuals have control over the pace in which they complete tasks. \cr
    control.breaks: \tab  A categorical variable to determine whether individuals have control over whether they amount of breaks they take in between completing tasks. \cr
    supervisor.support: \tab  A categorical variable to determine whether individuals feel they have support from their supervisor. \cr
    coworker.support: \tab  A categorical variable to determine whether individuals feel they have support from their coworkers. \cr
    job.satisfied: \tab  A categorical variable to determine whether individuals feel satisfied with their current job. \cr
    bmi: \tab The calculated body mass index (BMI) of individuals based on \code{height} and \code{weight}. \cr

  }
}

