% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsd-package.R
\docType{package}
\name{wlsd-package}
\alias{wlsd}
\alias{wlsd-package}
\title{wlsd: Wrangling Longitudinal Survival Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Supports the conversion of data between formats commonly used in longitudinal survival analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ci2131a/wlsd}
  \item Report bugs at \url{https://github.com/ci2131a/wlsd/issues}
}

}
\author{
\strong{Maintainer}: Charles Ingulli \email{charlesfi@outlook.com}

}
\keyword{internal}
